deal with some poorly documented quirks (#3)
X-Plane can be listening on a different UDP port than the default, if the user configured it this way. But thanks to the beacon, you will know which port the user set! So, you should not assume the X-Plane default port, but always use the one that the beacon packet tells you.

X-Plane will send at most 183 values in one RREF packet. That is because of the MTU of 1500 bytes on Ethernet. Subtract 20 bytes for the IP header, 8 bytes for the UDP header, and 5 bytes for the X-Plane header, you will have space for 183 values without fragmentation.

X-Plane will hold up to about 140 requests in the queue for one frame. So if you are requesting more than 140 RREFs, you should space the request out over more than one frame (50ms typically). I added code to wait after 100 requests.